Added store data in db

+++ scripts/tables/events.sql
@@ -2,7 +2,7 @@ CREATE TABLE events (
   id BIGINT NOT NULL AUTO_INCREMENT,
   device TINYINT NOT NULL,
   timestamp TIMESTAMP NOT NULL,
-  co2 INT,
+  co2 SMALLINT,
   temperature DOUBLE,
   humidity DOUBLE,
   PRIMARY KEY (id)
